+++ vhs-frontend/src/pages/Edit.tsx
@@ -1,8 +1,10 @@
 import axios from "axios";
 import { useState } from "react";
-import { VhsForm } from "../components/VhsForm";
-import { useLocation, useNavigate, redirect } from "react-router-dom";
+import { VhsForm } from "../components/form/VhsForm";
+import { useLocation, useNavigate } from "react-router-dom";
 import { Popup } from "../components/DeletePopup";
+import { createData } from "../components/form/CreateData";
+import { VhsFormType } from "../types";
 
 export const Edit = () => {
   const VHS = useLocation().state;
@@ -30,10 +32,9 @@ export const Edit = () => {
     }
   };
 
-  const handleSubmit = (data: any) => {
+  const handleSubmit = (data: VhsFormType) => {
     console.log(data);
-    editEntry(data);
-    redirect(`/${VHS.id}`);
+    editEntry(createData(data));
   };
 
   return (
